test: change test entity creation to domain
Change name of methods from entity to domain to show what it is doing
